% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2cat.R
\name{cat2cat_ml}
\alias{cat2cat_ml}
\title{The internal function used in the cat2cat ones}
\usage{
cat2cat_ml(ml, mapp, target_data, base_data, mid_data)
}
\arguments{
\item{ml}{`list` the same `ml` argument as provided to `cat2cat` function.}

\item{mapp}{`list` a mapping table}

\item{target_data}{`data.frame`}

\item{base_data}{`data.frame`}

\item{cat_mid}{`data.frame`}
}
\description{
apply the ml models to the cat2cat data
}
\keyword{internal}
