% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2cat.R
\name{dummy_c2c_cols}
\alias{dummy_c2c_cols}
\title{Add default cat2cat columns to a `data.frame`}
\usage{
dummy_c2c_cols(df, cat_var)
}
\arguments{
\item{df}{`data.frame`}

\item{cat_var}{`character` a categorical variable name}
}
\description{
a utils function to add default cat2cat columns to a `data.frame`.
It will be useful e.g. for a boarder periods which will not have additional `cat2cat` columns.
}
\examples{
data(occup_small)
data(occup)
data(trans)

occup_old <- occup_small[occup_small$year == 2008, ]
dummy_c2c_cols(occup_old, "code")
}
