% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occup}
\alias{occup}
\title{Occupational dataset}
\format{
A data frame with around 70000 observations and 12 variables.
\describe{
\item{id}{	integer	id}
\item{age}{ numeric age of a subject}
\item{sex}{ numeric sex of a subject}
\item{edu}{ integer edu level of education of a subject where lower means higher - 1 for at least master degree}
\item{exp}{ numeric exp number of experience years for a subject}
\item{district}{ integer district}
\item{parttime}{ numeric contract type regards time where 1 mean full-time (work a whole week)}
\item{salary}{ numeric salary per year}
\item{code}{ character code - occupational code}
\item{multiplier}{ numeric multiplier for the subject to reproduce a population - how many of such subjects in population}
\item{year}{integer year}
\item{code4}{ character code - occupational code - first 4 digits}
}
}
\usage{
occup
}
\description{
Occupational dataset
}
\details{
occup dataset is an example of unbalance panel dataset.
This is a simulated data although there are applied a real world characteristics from national statistical office survey.
The orginal survey is anonymous and take place every two years.
It is presenting a characteristics from randomly selected company and then using k step procedure employees are chosen.

occupational dataset
}
\keyword{datasets}
