% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2cat.R
\name{resolve_frequencies}
\alias{resolve_frequencies}
\title{Resolve the frequencies}
\usage{
resolve_frequencies(cat_base_year, cat_var_base, freqs_df, multiplier_var)
}
\arguments{
\item{cat_base_year}{`data.frame` a base period dataset.}

\item{cat_var_base}{`character(1)` name of the base period categorical variable.}

\item{freqs_df}{`data.frame` with 2 columns, the first one with category and second with counts.}

\item{multiplier_var}{`character(1)` name of the multiplier variable.}
}
\description{
Resolve the frequencies
}
\keyword{internal}
