% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trans}
\alias{trans}
\title{trans dataset containing mappings (transitions) between old (2008) and new (2010) occupational codes.
This table could be used to map encodings in both directions.}
\format{
A data frame with 2693 observations and 2 variables.
\describe{
\item{old}{	character an old encoding of a certain occupation}
\item{new}{	character a new encoding of a certain occupation}
}
}
\usage{
trans
}
\description{
trans dataset containing mappings (transitions) between old (2008) and new (2010) occupational codes.
This table could be used to map encodings in both directions.
}
\details{
mapping (transition) table for occupations where first column contains old encodings and second one a new encoding
}
\keyword{datasets}
