% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2cat.R
\name{validate_cover_cats}
\alias{validate_cover_cats}
\title{Validate if the trans table contains all proper mappings}
\usage{
validate_cover_cats(u_cats_target, mapp)
}
\arguments{
\item{u_cats_target}{vector of unique target period categories}

\item{mapp}{transition (mapping) table process with `get_mappings`,
the "to_base" direction is taken.}
}
\description{
Validate if the trans table contains all proper mappings
}
\keyword{internal}
