% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{verticals}
\alias{verticals}
\title{verticals dataset}
\format{
A data frame with 21 observations and 4 variables.
\describe{
\item{vertical}{	character an certain sales vertical}
\item{sales}{	numeric a size of sale}
\item{counts}{	integer counts size}
\item{v_date}{	character Date}
}
}
\usage{
verticals
}
\description{
verticals dataset
}
\details{
random data - aggregate sales across e-commerce verticals
}
\examples{
\dontrun{
agg_old <- data.frame(
 vertical = c("Electronics", "Kids1", "Kids2", "Automotive", "Books",
              "Clothes", "Home", "Fashion", "Health", "Sport"),
 sales = rnorm(10, 100, 10),
 counts = rgeom(10, 0.0001),
 v_date = rep("2020-04-01", 10), stringsAsFactors = F
)

agg_new <- data.frame(
 vertical = c("Electronics", "Supermarket", "Kids", "Automotive1",
              "Automotive2", "Books", "Clothes", "Home", "Fashion", "Health", "Sport"),
 sales = rnorm(11, 100, 10),
 counts = rgeom(11, 0.0001),
 v_date = rep("2020-05-01", 11), stringsAsFactors = F
)
verticals <- rbind(agg_old, agg_new)
}
}
\keyword{datasets}
