% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{verticals2}
\alias{verticals2}
\title{verticals2 dataset}
\format{
A data frame with 1000 observations and 4 variables.
\describe{
\item{ean}{ product ean}
\item{vertical}{	character an certain sales vertical}
\item{sales}{	numeric a size of sale}
\item{v_date}{	character Date}
}
}
\usage{
verticals2
}
\description{
verticals2 dataset
}
\details{
random data
}
\examples{
\dontrun{
vert_old <- data.frame(
 ean = 90000001:90000500,
 vertical = sample(c("Electronics", "Kids1", "Kids2", "Automotive", "Books",
              "Clothes", "Home", "Fashion", "Health", "Sport"), 500, replace = TRUE),
 sales = rnorm(500, 100, 10),
 v_date = rep("2020-04-01", 500), stringsAsFactors = F
)

vert_new <- vert_old
vert_new$v_date <- "2020-05-01"
vert_new$vertical[vert_new$vertical \%in\% c("Kids1", "Kids2")] <- "Kids"
vert_new$vertical[vert_new$vertical \%in\% c("Automotive")] <-
sample(c("Automotive1", "Automotive2"), sum(vert_new$vertical \%in\% c("Automotive")), replace = TRUE)
vert_new$vertical[vert_new$vertical \%in\% c("Home")] <-
sample(c("Home", "Supermarket"), sum(vert_new$vertical \%in\% c("Home")), replace = TRUE)

verticals2 <- rbind(vert_old, vert_new)
}
}
\keyword{datasets}
