% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2cat.R
\name{cat2cat}
\alias{cat2cat}
\title{Automatic transforming of a categorical variable according to a new encoding}
\usage{
cat2cat(
  data = list(old = NULL, new = NULL, cat_var = NULL, time_var = NULL, multipier_var =
    NULL),
  mappings = list(trans = NULL, direction = NULL),
  ml = list(method = NULL, features = NULL, args = NULL)
)
}
\arguments{
\item{data}{list with 4 or 5 named fileds `old` `new` `cat_var` `time_var` and optional `multipier_var`}

\item{mappings}{list with 2 named fileds `trans` `direction`}

\item{ml}{list with 3 named fields `method` `features` `args``}
}
\value{
list of 2 data.frames
}
\description{
Automatic transforming of a categorical variable according to a new encoding
}
\details{
data args
\itemize{
 \item{"old"}{ data.frame}
 \item{"new"} { data.frame}
 \item{"cat_var"}{ name of the caterogical variable}
 \item{"time_var"}{ name of the time varaiable}
 \item{"multipier_var"}{ name of the multipier variable - number of replication needed to reproduce the population}
}
mappings args
\itemize{
 \item{"trans"}{ data.frame with 2 columns - transition table}
 \item{"direction"}{ direction - "backward or "forward"}
}
ml args
\itemize{
 \item{"method"}{ character "knn"}
 \item{"features"}{ vector of features names}
 \item{"args"}{ k number of k nearest n}
}
}
\examples{
data(occup)
data(trans)

occup_old = occup[occup$year == 2008,]
occup_new = occup[occup$year == 2010,]

cat2cat(
  data = list(old = occup_old, new = occup_new, cat_var = "code", time_var = "year"),
  mappings = list(trans = trans, direction = "forward")
      )

cat2cat(
 data = list(old = occup_old ,new = occup_new, cat_var = "code", time_var = "year"),
 mappings = list(trans = trans, direction = "forward")
 ,ml = list(method = "knn", features = c("age", "sex", "edu", "exp", "parttime", "salary"), args = list(k = 10))
 )
}
