% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2cat_man.R
\name{cat2cat_man}
\alias{cat2cat_man}
\title{Manual transforming of a categorical variable according to a new encoding}
\usage{
cat2cat_man(
  data = list(old = NULL, new = NULL, cat_var = NULL, time_var = NULL, freq_var = NULL),
  ...
)
}
\arguments{
\item{data}{list with 5 named fileds `old` `new` `cat_var` `time_var` `freq_var`}

\item{...}{equations}
}
\description{
Manual transforming of a categorical variable according to a new encoding
}
\details{
data args
\itemize{
 \item{"old"}{ Stuff}
 \item{"new"}{ Stuff}
 \item{"cat_var"}{ Stuff}
 \item{"time_var"}{ Stuff}
 \item{"freq_var"}{ Stuff}
}
... equations
}
