% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2cat_ml.R
\name{cat2cat_ml}
\alias{cat2cat_ml}
\title{The internal function used in the cat2cat one}
\usage{
cat2cat_ml(ml, mapp, target_data, cat_var_target)
}
\arguments{
\item{ml}{`list` the same `ml` argument as provided to `cat2cat` function.}

\item{mapp}{`list` a mapping table}

\item{target_data}{`data.frame`}

\item{cat_var_target}{`character(1)` name of the categorical variable
in the target period.}
}
\description{
apply the ml models to the cat2cat data
}
\keyword{internal}
