% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2cat.R
\name{get_freqs}
\alias{get_freqs}
\title{getting frequencies for a vector with an optional multipier argument}
\usage{
get_freqs(x, multipier = NULL)
}
\arguments{
\item{x}{vector}

\item{multipier}{vector how many times to repeat certain value}
}
\value{
data.frame with two columns `input` `Freq`
}
\description{
getting frequencies for a vector with an optional multipier argument
}
\examples{
data(occup)

get_freqs(occup$code[occup$year == "2008"])
get_freqs(occup$code[occup$year == "2010"])

get_freqs(occup$code[occup$year == "2008"], occup$multipier[occup$year == "2008"])
get_freqs(occup$code[occup$year == "2010"], occup$multipier[occup$year == "2010"])
}
