% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2cat.R
\name{get_mappings}
\alias{get_mappings}
\title{transforming table of mappings to list with keys}
\usage{
get_mappings(x = data.frame())
}
\arguments{
\item{x}{data.frame with 2 columns}
}
\value{
list with 2 fields `to_old` `to_new`
}
\description{
transforming table of mappings to list with keys where first column is assumed to be an old encoding.
}
\examples{
data(trans)

mappings <- get_mappings(trans)
mappings$to_old[1:4]
mappings$to_new[1:4]
}
