% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lms.R
\name{summary_c2c}
\alias{summary_c2c}
\title{Summary for data.frame with replicated rows}
\usage{
summary_c2c(x, df_old, df_new)
}
\arguments{
\item{x}{lm object}

\item{df_old}{integer number of d.f in orginal dataset}

\item{df_new}{integer number of d.f in dataset with replicated rows}
}
\value{
data.frame
}
\description{
transforming sumary.lm object according to real number of d.f.
}
\examples{
data(occup)
data(trans)

occup_old <- occup[occup$year == 2008, ]
occup_new <- occup[occup$year == 2010, ]

occup_2 <- cat2cat(
  data = list(old = occup_old, new = occup_new, cat_var = "code", time_var = "year"),
  mappings = list(trans = trans, direction = "backward"),
  ml = list(method = "knn",
            features = c("age", "sex", "edu", "exp", "parttime", "salary"),
            args = list(k = 10))
)

# Regression
# we have to adjust size of std as we artificialy enlarge degrees of freedom
lms <- lm(I(log(salary)) ~ age + sex + factor(edu) + parttime + exp, occup_2$old,
          weights = multipier * wei_freq_c2c)
summary_c2c(lms, df_old = nrow(occup_old), df_new = nrow(occup_2$old))
}
